Create SQUEAKY CLEAN
This is an exercise titled SQUEAKY CLEAN.  It can be found on the Exercism.io website and has the objective of cleaning up various aspects of a String input, including replacing characters with other characters and omitting characters that should not be included.  
I found this to be a challenging exercise, in that the placement of curly braces was absolutely vital and taught me to pay close attention to them.